feat: avoid publlic folder
Moves assets out of the public folder. This is a special folder that directly exports things as-is by vite.

For our case however assets should just be referenced and used.

Additionally this fixes the imports for the other assets by importing and referencing them as expected.

This should reduce our build size if we don't use something.